Time measured Elapsed double ms replaced with int

+++ Sorter.Algorithms/EventArg/SortCompleteEventArgs.cs
@@ -4,17 +4,17 @@ namespace Sorter.Algorithms.EventArg
 {
     public sealed class SortCompleteEventArgs : EventArgs
     {
-        public double StartTimeMilliSec { get; private set; }
+        public int StartTimeMilliSec { get; private set; }
 
-        public double StopTimeMilliSec { get; private set; }
+        public int StopTimeMilliSec { get; private set; }
 
-        public double ElapsedTimeMilliSec { get; private set; }
+        public int ElapsedTimeMilliSec { get; private set; }
 
         public int ItemSortCount { get; private set; }
 
 
-        public SortCompleteEventArgs(double startTimeMilliSec, double stopTimeMilliSec, 
-                                     double elapsedTimeMilliSec, int itemSortCount)
+        public SortCompleteEventArgs(int startTimeMilliSec, int stopTimeMilliSec, 
+                                     int elapsedTimeMilliSec, int itemSortCount)
         {
             StartTimeMilliSec = startTimeMilliSec;
             StopTimeMilliSec = stopTimeMilliSec;
